refactor: querryHelper -> support, mapper로 분리
