+++ src/test/java/ssammudan/cotree/integration/DataJpaTestSupporter.java
@@ -14,7 +14,7 @@ import ssammudan.cotree.model.education.techbook.category.repository.TechBookEdu
 import ssammudan.cotree.model.education.techbook.techbook.repository.TechBookRepository;
 import ssammudan.cotree.model.education.techtube.category.repository.TechTubeEducationCategoryRepository;
 import ssammudan.cotree.model.education.techtube.techtube.repository.TechTubeRepository;
-import ssammudan.cotree.model.project.project.repository.helper.ProjectQueryHelper;
+import ssammudan.cotree.model.project.project.repository.support.ProjectQuerySupport;
 import ssammudan.cotree.model.review.review.repository.TechEducationReviewRepository;
 import ssammudan.cotree.model.review.reviewtype.repository.TechEducationTypeRepository;
 
@@ -70,7 +70,7 @@ public abstract class DataJpaTestSupporter {
 	protected TechTubeEducationCategoryRepository techTubeEducationCategoryRepository;
 
 	@MockitoBean
-	protected ProjectQueryHelper projectQueryHelper;
+	protected ProjectQuerySupport projectQuerySupport;
 
 	@BeforeEach
 	protected void setup() {
